requirements.txt
wheel >= 0.26.0
typing >=3.5.0.1; python_version < '3.5'
pyelftools >=0.23
